+++ src/Hyyan/WPI/Admin/Settings.php
@@ -102,10 +102,11 @@ class Settings extends \WeDevs_Settings_API
     {
         $options = get_option($section);
 
-        if (isset($options[$option])) {
-            return $options[$option];
-        }
-
-        return $default;
+        if (!empty($options[$option]))  // equivalent to: isset($options[$option]) && $options[$option]
+            return $options[$option];   // when all settings are disabled
+        elseif (isset($options[$option])) 
+            return array();
+        else
+            return $default;
     }
 }