+++ src/Hyyan/WPI/Cart.php
@@ -11,6 +11,7 @@
 namespace Hyyan\WPI;
 
 use Hyyan\WPI\Product\Variation;
+use Hyyan\WPI\Utilities;
 
 /**
  * Cart.
@@ -66,8 +67,8 @@ class Cart
         foreach (WC()->cart->get_cart() as $values) {
             $product = $values['data'];
 
-            if (in_array($product->id, $IDS)) {
-                $result = $product->id;
+            if (in_array($product->get_id(), $IDS)) {
+                $result = $product->get_id();
                 break;
             }
         }
@@ -91,7 +92,7 @@ class Cart
         // By default, returns the same input
         $cart_item_data_translation = $cart_item_data;
 
-        switch ( $cart_item_data->product_type ) {
+        switch ( $cart_item_data->get_type() ) {
             case 'variation':
                 $variation_translation   = $this->getVariationTranslation( $cart_variation_id );
                 $cart_item_data_translation = $variation_translation ? $variation_translation : $cart_item_data_translation;
@@ -265,7 +266,8 @@ class Cart
 
         // If no variation ID is set, attempt to get a variation ID from posted attributes.
         if (empty($variation_id)) {
-            $variation_id = $adding_to_cart->get_matching_variation(wp_unslash($_POST));
+            $data_store   = WC_Data_Store::load('product');
+			$variation_id = $data_store->find_matching_product_variation($adding_to_cart, wp_unslash($_POST));
         }
 
         /**
@@ -283,7 +285,7 @@ class Cart
 
             // Get the respective variation in the language of the product in the cart
             $variation    = $this->getVariationTranslation($variation_id, $lang);
-            $variation_id = $variation->variation_id;
+            $variation_id = $variation->get_id();
         } else {
             $variation = wc_get_product($variation_id);
         }
@@ -291,76 +293,93 @@ class Cart
          * End of custom code.
          */
 
-        //$variation = wc_get_product( $variation_id );
+        // Validate the attributes.
+		try {
+			if (empty($variation_id)) {
+				throw new Exception(__('Please choose product options&hellip;', 'woocommerce') );
+			}
 
-        // Verify all attributes
-        foreach ($attributes as $attribute) {
-            if (!$attribute['is_variation']) {
-                    continue;
-            }
+			$variation_data = wc_get_product_variation_attributes($variation_id);
 
-            $taxonomy = 'attribute_' . sanitize_title($attribute['name']);
+			foreach ($attributes as $attribute) {
+				if (!$attribute['is_variation']) {
+					continue;
+				}
 
-            if (isset($_REQUEST[$taxonomy])) {
-                // Get value from post data
-                if ($attribute['is_taxonomy']) {
-                    // Don't use wc_clean as it destroys sanitized characters
-                    $value = sanitize_title(stripslashes($_REQUEST[$taxonomy]));
+				$taxonomy = 'attribute_' . sanitize_title($attribute['name']);
 
-                    /**
-                    * Custom code to check if a translation of the product is already in the cart,
-                    * and in that case, replace the variation attribute being added to the cart by
-                    * the respective translation in the language of the product already in the cart
-                    * NOTE: The product_id is filtered by $this->add_to_cart() and holds the id of
-                    * the product translation, if one exists in the cart.
-                    */
-                    if ($product_id != absint($_REQUEST['add-to-cart'])) {
-                        // Get the translation of the term
-                        $term  = get_term_by('slug', $value, $attribute['name']);
-                        $_term = get_term_by('id', pll_get_term(absint($term->term_id), $lang), $attribute['name']);
-
-                        if ($_term) {
-                            $value = $_term->slug;
+				if (isset($_REQUEST[$taxonomy])) {
+					// Get value from post data
+					if ($attribute['is_taxonomy']) {
+						// Don't use wc_clean as it destroys sanitized characters
+						$value = sanitize_title(stripslashes($_REQUEST[$taxonomy]));
+                        
+                        /**
+                        * Custom code to check if a translation of the product is already in the cart,
+                        * and in that case, replace the variation attribute being added to the cart by
+                        * the respective translation in the language of the product already in the cart
+                        * NOTE: The product_id is filtered by $this->add_to_cart() and holds the id of
+                        * the product translation, if one exists in the cart.
+                        */
+                        if ($product_id != absint($_REQUEST['add-to-cart'])) {
+                            // Get the translation of the term
+                            $term  = get_term_by('slug', $value, $attribute['name']);
+                            $_term = get_term_by('id', pll_get_term(absint($term->term_id), $lang), $attribute['name']);
+    
+                            if ($_term) {
+                                $value = $_term->slug;
+                            }
                         }
-                    }
-                    /**
-                    * End of custom code.
-                    */
-                } else {
-                    $value = wc_clean(stripslashes($_REQUEST[$taxonomy]));
-                }
+                        /**
+                        * End of custom code.
+                        */
+					} else {
+						$value = wc_clean(stripslashes($_REQUEST[$taxonomy]));
+					}
 
-                // Get valid value from variation
-                $valid_value = isset($variation->variation_data[$taxonomy]) ? $variation->variation_data[$taxonomy] : '';
+					// Get valid value from variation
+//change proposed by @theleemon
+$variation_data = wc_get_product_variation_attributes($variation->get_id());
+$valid_value = isset($variation_data[$taxonomy]) ? $variation_data[$taxonomy] : '';
+					//$valid_value = isset($variation_data[$taxonomy]) ? $variation_data[$taxonomy] : '';
 
-                // Allow if valid
-                if ('' === $valid_value || $valid_value === $value) {
-                    $variations[$taxonomy] = $value;
-                    continue;
-                }
-            } else {
-                $missing_attributes[] = wc_attribute_label($attribute['name']);
-            }
-        }
+					// Allow if valid or show error.
+					if ( '' === $valid_value || $valid_value === $value) {
+						$variations[$taxonomy] = $value;
+					} else {
+						throw new Exception(sprintf(__('Invalid value posted for %s', 'woocommerce'), wc_attribute_label($attribute['name'])));
+					}
+				} else {
+					$missing_attributes[] = wc_attribute_label($attribute['name']);
+				}
+			}
+			if (!empty($missing_attributes)) {
+				throw new Exception(sprintf(_n('%s is a required field', '%s are required fields', sizeof($missing_attributes), 'woocommerce'), wc_format_list_of_items($missing_attributes)));
+			}
+		} catch (Exception $e) {
+			wc_add_notice($e->getMessage(), 'error');
+            //return false;
+            /**
+             * Custom code: We are doing an action, therefore no return needed. Instead we need to
+             * set $passed_validation to false, to ensure the product is not added to the card.
+             */
+            add_filter('woocommerce_add_to_cart_validation', '__return_false');
+		}
 
-        if (!empty($missing_attributes)) {
-            wc_add_notice(sprintf(_n('%s is a required field', '%s are required fields', sizeof($missing_attributes), 'woocommerce'), wc_format_list_of_items($missing_attributes)), 'error');
-        } elseif (empty($variation_id)) {
-            wc_add_notice(__('Please choose product options&hellip;', 'woocommerce'), 'error');
-        } else {
-            // Add to cart validation
-            $passed_validation 	= apply_filters('woocommerce_add_to_cart_validation', true, $product_id, $quantity, $variation_id, $variations);
+		// Add to cart validation
+		$passed_validation 	= apply_filters('woocommerce_add_to_cart_validation', true, $product_id, $quantity, $variation_id, $variations);
 
-            if ($passed_validation && WC()->cart->add_to_cart($product_id, $quantity, $variation_id, $variations) !== false) {
-                wc_add_to_cart_message(array($product_id => $quantity), true);
+		if ($passed_validation && WC()->cart->add_to_cart($product_id, $quantity, $variation_id, $variations) !== false) {
+			wc_add_to_cart_message(array($product_id => $quantity), true);
+			//return true;
+            /**
+             * Custom code: We are doing an action, therefore no return needed. Intead we need to
+             * set $was_added_to_cart to true to trigger the optional redirect
+             */
+            $was_added_to_cart = true;
+		}
 
-                //return true; Doing an action, no return needed but we need to set $was_added_to_cart to trigger the redirect
-                $was_added_to_cart = true;
-            } else {
-                $was_added_to_cart = false;
-            }
-        }
-        //return false; Doing an action, no return needed but we need to set $was_added_to_cart to trigger the redirect
+		//return false; We are doing an action, therefore no return needed. but we need to set $was_added_to_cart to trigger the redirect
         // End: From add_to_cart_handler_variable( $product_id )
 
         /**
@@ -401,8 +420,8 @@ class Cart
 
         // Get parent product translation id for the given language
         $variation   = wc_get_product($variation_id);
-        $parent      = $variation ? $variation->parent : null;
-        $_product_id = $parent ? pll_get_post($parent->id, $lang) : null;
+				$parentid = Utilities::get_variation_parentid($variation);
+				$_product_id = pll_get_post($parentid, $lang);
 
         // Get variation translation using the duplication metadata value
         $meta = get_post_meta($variation_id, Variation::DUPLICATE_KEY, true);