+++ src/Hyyan/WPI/Coupon.php
@@ -12,6 +12,7 @@ namespace Hyyan\WPI;
 
 use Hyyan\WPI\Admin\Settings;
 use Hyyan\WPI\Admin\Features;
+use Hyyan\WPI\Utilities;
 
 /**
  * Coupon.
@@ -41,11 +42,63 @@ class Coupon
      */
     public function couponLoaded(\WC_Coupon $coupon)
     {
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
+			  
         $productIDS = array();
         $excludeProductIDS = array();
         $productCategoriesIDS = array();
         $excludeProductCategoriesIDS = array();
 
+        foreach ($coupon->get_product_ids() as $id) {
+            foreach ($this->getProductPostTranslationIDS($id) as $_id) {
+                $productIDS[] = $_id;
+            }
+        }
+        foreach ($coupon->get_excluded_product_ids() as $id) {
+            foreach ($this->getProductPostTranslationIDS($id) as $_id) {
+                $excludeProductIDS[] = $_id;
+            }
+        }
+
+        foreach ($coupon->get_product_categories() as $id) {
+            foreach ($this->getProductTermTranslationIDS($id) as $_id) {
+                $productCategoriesIDS[] = $_id;
+            }
+        }
+
+        foreach ($coupon->get_excluded_product_categories() as $id) {
+            foreach ($this->getProductTermTranslationIDS($id) as $_id) {
+                $excludeProductCategoriesIDS[] = $_id;
+            }
+        }
+
+				$coupon->set_product_ids( $productIDS );
+				$coupon->set_excluded_product_ids( $excludeProductIDS );
+				$coupon->set_product_categories( $productCategoriesIDS );
+				$coupon->set_excluded_product_categories( $excludeProductCategoriesIDS );
+
+        return $coupon;
+    }
+				else
+				{
+					return $this->couponLoadedOld($coupon);
+				}
+		}
+
+    /**
+     * Extend the coupon to include porducts translations.
+     *
+     * @param \WC_Coupon $coupon
+     *
+     * @return \WC_Coupon
+     */
+    public function couponLoadedOld(\WC_Coupon $coupon)
+    {
+        $productIDS = array();
+        $excludeProductIDS = array();
+        $productCategoriesIDS = array();
+        $excludeProductCategoriesIDS = array();
         foreach ($coupon->product_ids as $id) {
             foreach ($this->getProductPostTranslationIDS($id) as $_id) {
                 $productIDS[] = $_id;
@@ -56,27 +109,22 @@ class Coupon
                 $excludeProductIDS[] = $_id;
             }
         }
-
         foreach ($coupon->product_categories as $id) {
             foreach ($this->getProductTermTranslationIDS($id) as $_id) {
                 $productCategoriesIDS[] = $_id;
             }
         }
-
         foreach ($coupon->exclude_product_categories as $id) {
             foreach ($this->getProductTermTranslationIDS($id) as $_id) {
                 $excludeProductCategoriesIDS[] = $_id;
             }
         }
-
         $coupon->product_ids = $productIDS;
         $coupon->exclude_product_ids = $excludeProductIDS;
         $coupon->product_categories = $productCategoriesIDS;
         $coupon->exclude_product_categories = $excludeProductCategoriesIDS;
-
         return $coupon;
     }
-
     /**
      * Get array of product translations IDS.
      *
@@ -89,7 +137,7 @@ class Coupon
         $result = array($ID);
         $product = wc_get_product($ID);
 
-        if ($product && $product->product_type === 'variation') {
+        if ($product && $product->get_type() === 'variation') {
             $IDS = Product\Variation::getRelatedVariation($ID, true);
             if (is_array($IDS)) {
                 $result = array_merge($result, $IDS);