+++ src/Hyyan/WPI/Product/Stock.php
@@ -52,9 +52,18 @@ class Stock
         $items = $order->get_items();
 
         /* Reduce stock */
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
         foreach ($items as $item) {
             $this->change($item, self::STOCK_REDUCE_ACTION);
         }
+    }
+				else
+				{
+					foreach ($items as $item) {
+							$this->change_old($item, self::STOCK_REDUCE_ACTION);
+					}
+				}		
     }
 
     /**
@@ -73,11 +82,20 @@ class Stock
         $items = $order->get_items();
 
         /* Increase stock */
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
         foreach ($items as $item) {
-            $item['change'] = $change;
+            $item->change = $change;
             $this->change($item, self::STOCK_INCREASE_ACTION);
         }
-
+				}
+				else
+				{
+					foreach ($items as $item) {
+							$item->change = $change;
+							$this->change_old($item, self::STOCK_INCREASE_ACTION);
+					}
+				}		
         return $change;
     }
 
@@ -87,13 +105,13 @@ class Stock
      * @param array  $item   the order data
      * @param string $action STOCK_REDUCE_ACTION | STOCK_INCREASE_ACTION
      */
-    protected function change(array $item, $action = self::STOCK_REDUCE_ACTION)
+protected function change( \WC_Order_Item_Product $item, $action = self::STOCK_REDUCE_ACTION )
     {
-        $productID = $item['product_id'];
+				$productID = Utilities::get_order_item_productid($item);
         $productObject = wc_get_product($productID);
         $productLang = pll_get_post_language($productID);
 
-        $variationID = $item['variation_id'];
+        $variationID = Utilities::get_order_item_variationid($item);
 
         /* Handle Products */
         if ($productObject && $productLang) {
@@ -109,11 +127,11 @@ class Stock
             $isManageStock = $productObject->managing_stock();
             $isVariation = $variationID && $variationID > 0;
             $method = ($action === self::STOCK_REDUCE_ACTION) ?
-                    'reduce_stock' :
-                    'increase_stock';
+                    'decrease' :
+                    'increase';
             $change = ($action === self::STOCK_REDUCE_ACTION) ?
-                    $item['qty'] :
-                    $item['change'];
+              Utilities::get_order_item_quantity($item) :
+          		Utilities::get_order_item_change($item);
 
             /* Sync stock for all translation */
             foreach ($translations as $ID) {
@@ -121,7 +139,7 @@ class Stock
                 /* Only if product is managing stock */
                 if ($isManageStock) {
                     if (($translation = wc_get_product($ID))) {
-                        $translation->$method($change);
+                        \wc_update_product_stock($translation, $change, $method);
                     }
                 }
 
@@ -135,8 +153,72 @@ class Stock
                 }
             }
 
-            /* Handle variation */
-            if ($isVariation) {
+            /* Handle variation stock UNLESS stock is managed on the parent
+						 * there is a function for this $variation->get_stock_managed_by_id() however in woo-poly-context 
+						 * this returns the master language id of either the variation of the parent.
+						 */
+            if (($isVariation) && !($isManageStock)) {
+                $posts = Variation::getRelatedVariation($variationID);
+                foreach ($posts as $post) {
+                    if ($post->ID == $variationID) {
+                        continue;
+                    }
+                    $variation = wc_get_product($post);
+                    if ($variation && $variation->managing_stock()) {
+												\wc_update_product_stock($variation, $change, $method);
+                    }
+                }
+            }
+        }
+    }
+		
+		/** OLD WOO < 2-6 version
+     * Change the product stock in the given order item.
+     *
+     * @param array  $item   the order data
+     * @param string $action STOCK_REDUCE_ACTION | STOCK_INCREASE_ACTION
+     */
+    protected function change_old(array $item, $action = self::STOCK_REDUCE_ACTION)
+    {
+        $productID = $item['product_id'];
+        $productObject = wc_get_product($productID);
+        $productLang = pll_get_post_language($productID);
+        $variationID = $item['variation_id'];
+        /* Handle Products */
+        if ($productObject && $productLang) {
+            /* Get the translations IDS */
+            $translations = Utilities::getProductTranslationsArrayByObject(
+                            $productObject
+            );
+            /* Remove the current product from translation array */
+            unset($translations[$productLang]);
+            $isManageStock = $productObject->managing_stock();
+            $isVariation = $variationID && $variationID > 0;
+            $method = ($action === self::STOCK_REDUCE_ACTION) ?
+                    'reduce_stock' :
+                    'increase_stock';
+            $change = ($action === self::STOCK_REDUCE_ACTION) ?
+                    $item['qty'] :
+                    $item['change'];
+            /* Sync stock for all translation */
+            foreach ($translations as $ID) {
+                /* Only if product is managing stock */
+                if ($isManageStock) {
+                    if (($translation = wc_get_product($ID))) {
+                        $translation->$method($change);
+                    }
+                }
+                $general = Settings::getOption(
+                                'general', MetasList::getID(), array('total_sales')
+                );
+                if (in_array('total_sales', $general)) {
+                    update_post_meta(
+                            $ID, 'total_sales', get_post_meta($productID, 'total_sales', true)
+                    );
+                }
+            }
+            /* Handle variation stock UNLESS stock is managed on the parent */
+            if (($isVariation) && !($isManageStock)) {
                 $posts = Variation::getRelatedVariation($variationID);
                 foreach ($posts as $post) {
                     if ($post->ID == $variationID) {
@@ -150,4 +232,5 @@ class Stock
             }
         }
     }
+
 }