+++ src/Hyyan/WPI/Utilities.php
@@ -54,7 +54,7 @@ final class Utilities
      */
     public static function getProductTranslationsArrayByObject(\WC_Product $product, $excludeDefault = false)
     {
-        return static::getProductTranslationsArrayByID($product->id, $excludeDefault);
+        return static::getProductTranslationsArrayByID($product->get_id(), $excludeDefault);
     }
 
     /**
@@ -86,7 +86,7 @@ final class Utilities
      */
     public static function getProductTranslationByObject(\WC_Product $product, $slug = '')
     {
-        $productTranslationID = pll_get_post($product->id, $slug);
+        $productTranslationID = pll_get_post($product->get_id(), $slug);
 
         if ($productTranslationID) {
             $translated = wc_get_product($productTranslationID);
@@ -255,8 +255,15 @@ final class Utilities
         $product = wc_get_product( $product_id );
         $translated_attributes = array();
 
-        if ($product && 'variable' === $product->product_type) {
-            $default_attributes = $product->get_variation_default_attributes();
+        if ($product && 'variable' === $product->get_type()) {
+						if (Utilities::woocommerceVersionCheck('3.0')) 
+						{
+            $default_attributes = $product->get_default_attributes();
+						}
+						else
+						{
+							$default_attributes = $product->get_variation_default_attributes();
+						}
             $terms = array(); // Array of terms: if the term is taxonomy each value is a term object, otherwise an array (term slug => term value)
             $langs = array();
 
@@ -313,15 +320,15 @@ final class Utilities
         if (is_numeric($product))
             $product = wc_get_product(asbint($product));
 
-        if ($product && 'variable' === $product->product_type)
+        if ($product && 'variable' === $product->get_type())
             return true;
-        elseif ($product && 'simple' === $product->product_type) {
+        elseif ($product && 'simple' === $product->get_type()) {
             $current_screen  = function_exists('get_current_screen') ? get_current_screen() : false;
             $add_new_product = $current_screen && $current_screen->post_type === 'product' && $current_screen->action === 'add';
             $is_translation  = isset($_GET['from_post']) && isset($_GET['new_lang']);
             $has_variations  = get_children(array(
                     'post_type'   => 'product_variation',
-                    'post_parent' => $product->id
+                    'post_parent' => $product->get_id()
                 ));
 
             if ($add_new_product && $is_translation && $has_variations)
@@ -330,4 +337,160 @@ final class Utilities
 
         return false;
     }
+		
+    /**
+     * get payment method for order independent of wooCommerce version
+     *
+     * @param WC_Order $order
+     *
+     * @return string payment method name.
+     */
+    public static function get_payment_method($order)
+    {		
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
+					return $order->get_payment_method();
+				}
+				else
+				{
+					return $order->payment_method;
+				}
+		}
+		/**
+     * get billing country for order independent of wooCommerce version
+     *
+     * @param WC_Order $order
+     *
+     * @return string payment method name.
+     */
+    public static function get_billing_country($order)
+    {		
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
+					return $order->get_billing_country();
+				}
+				else
+				{
+					return $order->billing_country;
+				}
+		}
+
+		/**
+     * get product id for order item independent of wooCommerce version
+     *
+     * @param WC_Order_Item_Product $item
+     *
+     * @return id
+     */
+    public static function get_order_item_productid($item)
+    {		
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
+	        return $item->get_product_id();
+	  		}
+				else
+				{
+					return $item['product_id'];
+				}
+		}
+
+		
+		/**
+     * get variation id for order item independent of wooCommerce version
+     *
+     * @param WC_Order_Item_Product $item
+     *
+     * @return id
+     */
+    public static function get_order_item_variationid($item)
+    {		
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
+	        return $item->get_variation_id();
+	  		}
+				else
+				{
+					return $item['variation_id'];
+				}
+		}
+
+				
+		/**
+     * get quantity for order item independent of wooCommerce version
+     *
+     * @param WC_Order_Item_Product $item
+     *
+     * @return integer quantity
+     */
+    public static function get_order_item_quantity($item)
+    {		
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
+	        return $item->get_quantity();
+	  		}
+				else
+				{
+					return $item['qty'];
+				}
+		}
+
+		/**
+     * get change for order item independent of wooCommerce version
+     *
+     * @param WC_Order_Item_Product $item
+     *
+     * @return integer change
+     */
+    public static function get_order_item_change($item)
+    {		
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
+	        return $item->change;
+	  		}
+				else
+				{
+					return $item['change'];
+				}
+		}
+
+		/**
+     * get order languate independent of wooCommerce version
+     *
+     * @param WC_Order order
+     *
+     * @return string language
+     */
+    public static function get_orderid($order)
+    {		
+				// Get order language
+				if (Utilities::woocommerceVersionCheck('3.0')) 
+				{
+	        return $order->get_id();
+				}
+				else
+				{
+	        return $order->id;
+				}
+				
+		}
+
+
+
+		/**
+     * get id for variation parent independent of wooCommerce version
+     *
+     * @param WC_Product variation
+     *
+     * @return integer id of variation parent post
+     */
+    public static function get_variation_parentid($variation)
+    {		
+				if ($variation){
+					return (Utilities::woocommerceVersionCheck('3.0')) ? $variation->get_parent_id() : $variation->parent->get_id();
+				}
+				else
+				{
+					return null;
+				}
+		}
 }