+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 0.30.0
+
+* Fix #137 #131 #130 #110 adaptations for wooCommerce 3.0
+* Fix #136 Variable product stock sync issue where stock managed at parent level
+* Enh #132 Add settings Page link to plugins page
+* Fix #128 Allow variation description to be editable in translations
+* Fix #129 #138 Account page only shows orders in current language
+* Fix #112 Shipping Class are not sync for Product Variations
+
 ### 0.29.1
 
 * Improve Multisite compatibility